// betforge-front-fixed/src/types/api.ts
export interface WithdrawPayload {
  amount: number;
  method: 'pix' | 'ted' | 'crypto';
  pixKey?: string;
  pixKeyType?: 'cpf' | 'email' | 'phone' | 'random_key';
  bankData?: {
    bank: string;
    agency: string;
    account: string;
    accountType: 'checking' | 'savings';
  };
}